move field utils into one file
