+++ src/main/java/frc/robot/utils/autoaim/AutoAim.java
@@ -18,6 +18,9 @@ import edu.wpi.first.wpilibj2.command.Commands;
 import frc.robot.Robot;
 import frc.robot.Robot.RobotType;
 import frc.robot.subsystems.swerve.SwerveSubsystem;
+import frc.robot.utils.FieldUtils.AlgaeIntakeTargets;
+import frc.robot.utils.FieldUtils.CoralTargets;
+
 import java.util.List;
 import java.util.function.DoubleSupplier;
 import java.util.function.Supplier;
